MDE/PKFE-32 restructured toolbar button data
